Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -3,7 +3,7 @@
 int main() {
     char a;
     scanf("%c",&a);
-    int b= (int)a;
+    int b= a;
     printf("ASCII Value: %d",b);
     return 0;
 }
